moved base event classes to evetns folder. removed simulation fault class. removed few errors.
