+++ gov.pnnl.goss.gridappsd/src/gov/pnnl/goss/gridappsd/dto/TestConfig.java
@@ -147,12 +147,12 @@ public class TestConfig implements Serializable {
 		Fault fail = new Fault();
 		fail.event_type = Fault.class.getSimpleName();
 		fail.ObjectMRID = "235242342342342";
-		Map<FaultImpedance,Double> faultImpedanceMap = new HashMap<FaultImpedance, Double>();
-		faultImpedanceMap.put(FaultImpedance.rGround, 0.3);
-		faultImpedanceMap.put(FaultImpedance.xGround, 0.2);
-		fail.PhaseConnectedFaultKind = PhaseConnectedFaultKind.lineToGround;
+		Map<PhaseConnectedFaultKind,Double> faultKindMap = new HashMap<PhaseConnectedFaultKind, Double>();
+		faultKindMap.put(PhaseConnectedFaultKind.lineToGround, 0.3);
+		faultKindMap.put(PhaseConnectedFaultKind.lineToLineToGround, 0.2);
+		fail.PhaseConnectedFaultKind = faultKindMap;
 		fail.phases = PhaseCode.ABC;
-		fail.FaultImpedance = faultImpedanceMap;
+		fail.FaultImpedance = FaultImpedance.rGround;
 		fail.occuredDateTime = new Date().getTime();
 		fail.stopDateTime = new Date().getTime();
